+++ src/term/mod.rs
@@ -838,11 +838,16 @@ impl SizeInfo {
         Column(((self.width - 2. * self.padding_x) / self.cell_width) as usize)
     }
 
-    pub fn contains_point(&self, x: usize, y:usize) -> bool {
-        x < self.width as usize
-            && x >= 0 as usize
+    pub fn contains_point(&self, x: usize, y: usize, include_padding: bool) -> bool {
+        if include_padding {
+            x < self.width as usize
             && y < self.height as usize
-            && y >= 0 as usize
+        } else {
+            x < (self.width - self.padding_x) as usize
+                && x >= self.padding_x as usize
+                && y < (self.height - self.padding_y) as usize
+                && y >= self.padding_y as usize
+        }
     }
 
     pub fn pixels_to_coords(&self, x: usize, y: usize) -> Point {
@@ -1097,6 +1102,21 @@ impl Term {
 
     pub(crate) fn visible_to_buffer(&self, point: Point) -> Point<usize> {
         self.grid.visible_to_buffer(point)
+    }
+
+    /// Convert the given pixel values to a grid coordinate
+    ///
+    /// The mouse coordinates are expected to be relative to the top left. The
+    /// line and column returned are also relative to the top left.
+    ///
+    /// Returns None if the coordinates are outside the window,
+    /// padding pixels are considered inside the window
+    pub fn pixels_to_coords(&self, x: usize, y: usize) -> Option<Point> {
+        if self.size_info.contains_point(x, y, true) {
+            Some(self.size_info.pixels_to_coords(x, y))
+        } else {
+            None
+        }
     }
 
     /// Access to the raw grid data structure