+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Cursor jumping around when leaving alt screen while not in the alt screen
 - Text lingering around when resetting while scrolled up in the history
 - Terminfo support for extended capabilities
+- Mouse presses are no long ignored in padding except for event propagation
 
 ## Version 0.2.9
 
@@ -249,7 +250,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Empty lines in selections are now properly copied to the clipboard
 - Selection start point lagging behind initial cursor position
 - Rendering of selections which start above the visible area and end below it
-- Mouse press no longer ignored in padding width
 
 ### Deprecated